asmd::build - new build script

+++ build.sh
@@ -3,18 +3,12 @@
 # Advanced SmartOS Management Daemon
 # 2015 By Jorge Schrauwen
 # ----------------------------------
-# build an esky package
 #
-rm -rf dist/
-python setup.py sdist
-python setup.py bdist
-python setup.py bdist_esky
-rm -rf asmd.egg-info/ build/
-rm dist/*.gz
-mkdir -p dist/asmd/bin
-unzip dist/asmd-*.zip -d dist/asmd/bin
-rm dist/*.zip
-cp -r data/* dist/asmd/
-find dist/asmd/modules -type f -name "*.pyc" -exec rm {} \+
-tar czvf dist/asmd-$(grep version setup.py | awk -F'"' '{ print $2 }').tar.gz -C dist/ asmd/
-rm -rf dist/asmd/
+VERSION=1.0.0
+[ -d dist/ ] && rm -rf dist/
+mkdir -p dist/asmd
+mkdir -p dist/asmd/{bin,etc}
+cp asmd.sh dist/asmd/bin/asmd
+cp -r share dist/asmd/
+cp -r services dist/asmd/
+tar czvf dist/asmd-${VERSION}.tar.gz -C dist/ asmd/
